fix(review): align lifecycle helper with process facade
