+++ src/autoskillit/execution/process/__init__.py
@@ -14,7 +14,7 @@ from __future__ import annotations
 import functools
 import subprocess
 import time
-from collections.abc import Callable, Mapping, Sequence
+from collections.abc import Callable, Mapping
 from dataclasses import asdict
 from pathlib import Path
 from typing import TYPE_CHECKING, assert_never, cast
@@ -48,7 +48,6 @@ from autoskillit.execution.process._process_jsonl import (
     _jsonl_last_record_type,
     _marker_is_standalone,
     fold_event_cursor,
-    fold_event_lines,
 )
 from autoskillit.execution.process._process_kill import (
     _wait_process_dead,
@@ -73,6 +72,7 @@ from autoskillit.execution.process._process_race import (
     _watch_process,
     _watch_session_log,
     _watch_stdout_idle,
+    fold_lifecycle_evidence,
     resolve_termination,
 )
 
@@ -123,17 +123,6 @@ __all__ = [
     "run_managed_sync",
     "summarize_capture",
 ]
-
-
-def fold_lifecycle_evidence(
-    lines: Sequence[str],
-    parser: StreamParser,
-) -> tuple[tuple[str, ...], bool]:
-    """Reduce lifecycle records to pending-task and wakeup evidence."""
-    accumulator = RaceAccumulator(lifecycle_observation_enabled=True)
-    fold_event_lines(list(lines), parser, accumulator.observe_event)
-    signals = accumulator.to_race_signals()
-    return signals.pending_task_ids, signals.schedule_wakeup_violation
 
 
 def _resolve_session_id(
